WIP: PUT-login endpoint
