+++ api/api.rest
@@ -13,3 +13,22 @@ Content-Type: application/json
 GET http://localhost:9000/contents
 
 ###
+
+PUT http://localhost:9000/login
+Content-Type: application/json
+
+{
+  "username": "bob",
+  "password": "123"
+}
+
+###
+
+PUT http://localhost:9000/login
+Content-Type: application/json
+
+{
+  "username": "bob"
+}
+
+###
